Added Metal Type
Whew, this one was a doozy! I worked really hard on replacing the game's bit operations it used for a lot of type-based logic, only to work out a way I can extend them to work with more than 8 types and then deleted all that work LOL. I've done a fair amount of testing and the new type logic should be working as expected (outside of type-changing abilities, which is noted in TODO.md)

- Added Metal type which can be applied to Pokémon and Energy
- Added Basic Metal Energy (no special energy as of yet)
- Altered graphics files so that Metal now loads in with the other types
- Deck construction menu has been altered to account for Metal type
- AI should be able to account for Metal type but this hasn't been tested thoroughly beyond ensuring they can use Metal Energy to meet move energy costs
- Energy macro has been altered to account for more than 8 energy types
- Fixed the tutorial so that it properly verifies your moves - I commented some things out for testing and accidentally committed that
- Updated README and TODO based off new changes

Next step will be Fairy and Dragon types, and then another overhaul of the Deck Construction menu.